압축/비압축 이미지 비교, 밉맵 렌더2D 빌드

+++ Framework/Utilities/Math.cpp
@@ -46,10 +46,10 @@ Quaternion Math::ToQuaternion(Vector3 vec)
 	double sr = sin(roll * 0.5);
 
 	Quaternion q;
-	q.w = cr * cp * cy + sr * sp * sy;
-	q.x = sr * cp * cy - cr * sp * sy;
-	q.y = cr * sp * cy + sr * cp * sy;
-	q.z = cr * cp * sy - sr * sp * cy;
+	q.w = (FLOAT)(cr * cp * cy + sr * sp * sy);
+	q.x = (FLOAT)(sr * cp * cy - cr * sp * sy);
+	q.y = (FLOAT)(cr * sp * cy + sr * cp * sy);
+	q.z = (FLOAT)(cr * cp * sy - sr * sp * cy);
 
 	return q;
 }
@@ -60,19 +60,19 @@ Vector3 Math::ToEuler(Quaternion q)
 	// roll
 	double sinr_cosp = 2 * (q.w * q.x + q.y * q.z);
 	double cosr_cosp = 1 - 2 * (q.x * q.x + q.y * q.y);
-	angles.z = std::atan2(sinr_cosp, cosr_cosp);
+	angles.z = (float)std::atan2(sinr_cosp, cosr_cosp);
 	
 	//pitch
 	double sinp = 2 * (q.w * q.y - q.z * q.x);
 	if (abs(sinp) >= 1)
-		angles.x = copysign(Math::PI / 2.0, sinp); // use 90 degrees if out of range
+		angles.x = (float)copysign(Math::PI / 2.0, sinp); // use 90 degrees if out of range
 	else
-		angles.x = asin(sinp);
+		angles.x = (float)asin(sinp);
 
 	// yaw
 	double siny_cosp = 2 * (q.w * q.z + q.x * q.y);
 	double cosy_cosp = 1 - 2 * (q.y * q.y + q.z * q.z);
-	angles.y = atan2(siny_cosp, cosy_cosp);
+	angles.y = (float)atan2(siny_cosp, cosy_cosp);
 
 	return angles;
 }
